collectors/requirements.txt
requests>=2.31.0
msal>=1.26.0
python-dateutil>=2.9.0
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0
